+++ packages/react-components/src/Workers/useWorkers.ts
@@ -9,21 +9,27 @@ import { WorkersContext } from './workersContext';
 
 type UseWorkersWithoutId = () => {
   getWorker: (workerId: string) => GetWorkerOutput | undefined;
-  setWorker: (input: SetWorkerInput) => void;
+  setWorker: (input: SetWorkerInput) => Worker | undefined;
   removeWorker: (workerId: string) => void;
+  terminateWorker: (workerId: string) => void;
 };
 
 type UseWorkersWithId = (useWorkerWorkerId: string) => {
   getWorker: () => GetWorkerOutput | undefined;
-  setWorker: (input: SetWorkerWithWorkerIdInput) => void;
+  setWorker: (input: SetWorkerWithWorkerIdInput) => Worker | undefined;
   removeWorker: () => void;
+  terminateWorker: () => void;
 };
 
 const useWorkers: UseWorkersWithoutId & UseWorkersWithId = (
   useWorkerWorkerId?: string,
 ) => {
-  const { getWorker, setWorker, removeWorker }: WorkersContextInterface =
-    useContext(WorkersContext);
+  const {
+    getWorker,
+    setWorker,
+    removeWorker,
+    terminateWorker,
+  }: WorkersContextInterface = useContext(WorkersContext);
 
   const getWorkerByWorkerId = (workerId?: string) => {
     if (useWorkerWorkerId) {
@@ -39,14 +45,20 @@ const useWorkers: UseWorkersWithoutId & UseWorkersWithId = (
     workerId,
     worker,
     methods,
+    eventListeners,
   }: SetWorkerWithWorkerIdInput) => {
     if (useWorkerWorkerId) {
-      return setWorker({ workerId: useWorkerWorkerId, worker, methods });
+      return setWorker({
+        workerId: useWorkerWorkerId,
+        worker,
+        methods,
+        eventListeners,
+      });
     } else {
       if (!workerId) {
         throw new Error('workerId is required');
       }
-      return setWorker({ workerId, worker, methods });
+      return setWorker({ workerId, worker, methods, eventListeners });
     }
   };
 
@@ -61,10 +73,22 @@ const useWorkers: UseWorkersWithoutId & UseWorkersWithId = (
     }
   };
 
+  const terminateWorkerByWorkerId = (workerId?: string) => {
+    if (useWorkerWorkerId) {
+      terminateWorker(useWorkerWorkerId);
+    } else {
+      if (!workerId) {
+        throw new Error('workerId is required');
+      }
+      terminateWorker(workerId);
+    }
+  };
+
   return {
     getWorker: getWorkerByWorkerId,
     setWorker: setWorkerByWorkerId,
     removeWorker: removeWorkerByWorkerId,
+    terminateWorker: terminateWorkerByWorkerId,
   };
 };