+++ packages/react-components/src/dayjs/dayjs.tsx
@@ -1,29 +1,31 @@
-import dayjsInstance from 'dayjs'
-import esLocale from 'dayjs/locale/es'
-import customParseFormatPlugin from 'dayjs/plugin/customParseFormat.js'
-import dayOfYearPlugin from 'dayjs/plugin/dayOfYear.js'
-import durationPlugin from 'dayjs/plugin/duration.js'
-import isTodayPlugin from 'dayjs/plugin/isToday.js'
-import localeDataPlugin from 'dayjs/plugin/localeData.js'
-import relativeTimePlugin from 'dayjs/plugin/relativeTime.js'
-import timezonePlugin from 'dayjs/plugin/timezone.js'
-import toObjectPlugin from 'dayjs/plugin/toObject.js'
-import utcPlugin from 'dayjs/plugin/utc.js'
-import weekdayPlugin from 'dayjs/plugin/weekday.js'
+import dayjsInstance, { Dayjs } from 'dayjs'
+import esLocaleDefault from 'dayjs/locale/es'
+import customParseFormat from 'dayjs/plugin/customParseFormat.js'
+import dayOfYear from 'dayjs/plugin/dayOfYear.js'
+import duration from 'dayjs/plugin/duration.js'
+import isToday from 'dayjs/plugin/isToday.js'
+import localeData from 'dayjs/plugin/localeData.js'
+import relativeTime from 'dayjs/plugin/relativeTime.js'
+import timezone from 'dayjs/plugin/timezone.js'
+import toObject from 'dayjs/plugin/toObject.js'
+import utc from 'dayjs/plugin/utc.js'
+import weekday from 'dayjs/plugin/weekday.js'
 
-dayjsInstance.extend(utcPlugin)
-dayjsInstance.extend(relativeTimePlugin)
-dayjsInstance.extend(timezonePlugin)
-dayjsInstance.extend(localeDataPlugin)
-dayjsInstance.extend(dayOfYearPlugin)
-dayjsInstance.extend(isTodayPlugin)
-dayjsInstance.extend(weekdayPlugin)
-dayjsInstance.extend(toObjectPlugin)
-dayjsInstance.extend(durationPlugin)
-dayjsInstance.extend(customParseFormatPlugin)
-dayjsInstance.locale({...esLocale, weekStart: 1})
+const esLocale = {...esLocaleDefault, weekStart: 1}
 
-const getDayjs = (...args: any[]) =>  dayjsInstance(...args)
+dayjsInstance.extend(utc)
+dayjsInstance.extend(relativeTime)
+dayjsInstance.extend(timezone)
+dayjsInstance.extend(localeData)
+dayjsInstance.extend(dayOfYear)
+dayjsInstance.extend(isToday)
+dayjsInstance.extend(weekday)
+dayjsInstance.extend(toObject)
+dayjsInstance.extend(duration)
+dayjsInstance.extend(customParseFormat)
+dayjsInstance.locale(esLocale)
+
+const getDayjs = (...args: any[]): Dayjs =>  dayjsInstance(...args)
 
 getDayjs.weekdays = dayjsInstance.weekdays
 getDayjs.duration = dayjsInstance.duration
@@ -35,4 +37,19 @@ export const setDayjs = function (options: any = {}) {
   }
 }
 
-export const dayjs = (...args: any[]) => getDayjs(...args)
+export const dayjs = (...args: any[]): Dayjs => getDayjs(...args)
+
+export {
+  Dayjs,
+  utc,
+  relativeTime,
+  timezone,
+  localeData,
+  dayOfYear,
+  isToday,
+  weekday,
+  toObject,
+  duration,
+  customParseFormat,
+  esLocale
+}