+++ packages/react-components/src/dayjs/index.ts
@@ -1,4 +1,36 @@
-import { dayjs, setDayjs } from './dayjs';
+import {
+  dayjs, 
+  setDayjs, 
+  Dayjs, 
+  utc,
+  relativeTime,
+  timezone,
+  localeData,
+  dayOfYear,
+  isToday,
+  weekday,
+  toObject,
+  duration,
+  customParseFormat,
+  esLocale
+} from './dayjs';
 import { DayjsContext, DayjsProvider } from "./dayjsContext"
 
-export {DayjsContext, DayjsProvider, dayjs, setDayjs}
+export {
+  DayjsContext, 
+  DayjsProvider, 
+  dayjs, 
+  setDayjs, 
+  Dayjs, 
+  utc,
+  relativeTime,
+  timezone,
+  localeData,
+  dayOfYear,
+  isToday,
+  weekday,
+  toObject,
+  duration,
+  customParseFormat,
+  esLocale
+}